pyproject.toml
[project]
name = "logos-api"
version = "0.1.0"
description = "LogosAI FastAPI Backend Server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # FastAPI & Server
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sse-starlette>=2.0.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",

    # Validation & Settings
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",

    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.26.0",

    # AI/ML
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "aiohttp>=3.9.0",

    # RAG - ElasticSearch
    "elasticsearch>=8.12.0",
    "langchain-elasticsearch>=0.2.0",
    "langchain-huggingface>=0.0.3",
    "langchain-text-splitters>=0.2.0",
    "langchain-core>=0.2.0",

    # Document Processing
    "pymupdf>=1.23.0",
    "python-docx>=1.0.0",
    "aiofiles>=23.2.0",
    "PyPDF2>=3.0.0",
    "Pillow>=10.0.0",

    # ML Utils
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "scipy>=1.12.0",

    # Utilities
    "python-multipart>=0.0.6",
    "email-validator>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
